requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-mysql==5.2.0
mysql-connector-python==8.1.0